add unit tests for config_commands and couchdb_backup

+++ brewblox_ctl_lib/migrate.py
@@ -2,13 +2,19 @@
 Migration scripts
 """
 
-import sys
 from distutils.version import StrictVersion
-from os import getenv
+from os import getenv as getenv_
 
 from brewblox_ctl.commands import Command
-from brewblox_ctl.utils import check_config, confirm
-from brewblox_ctl_lib.const import CURRENT_VERSION, HISTORY
+from brewblox_ctl.utils import check_config, select
+
+from brewblox_ctl_lib.const import (CFG_VERSION_KEY, CURRENT_VERSION,
+                                    HISTORY_URL, PY)
+
+
+def getenv(env):  # pragma: no cover
+    """Wrapper function to improve testability"""
+    return getenv_(env)
 
 
 class MigrateCommand(Command):
@@ -23,9 +29,8 @@ class MigrateCommand(Command):
         if self.prev_version < StrictVersion('0.2.0'):
             # Breaking changes: Influx downsampling model overhaul
             # Old data is completely incompatible
-            if confirm('Upgrading to version >=0.2.0 requires a complete reset of your history data. ' +
-                       'Do you want to copy the data to a different directory (./influxdb-old) first?'):
-                shell_commands += ['sudo cp -rf ./influxdb ./influxdb-old']
+            select('Upgrading to version >=0.2.0 requires a complete reset of your history data. ' +
+                   'We\'ll be deleting it now')
             shell_commands += ['sudo rm -rf ./influxdb']
 
         return shell_commands
@@ -38,15 +43,15 @@ class MigrateCommand(Command):
             # Breaking changes: Influx downsampling model overhaul
             # Old data is completely incompatible
             shell_commands += [
-                'curl -Sk -X GET --retry 60 --retry-delay 10 {}/_service/status > /dev/null'.format(HISTORY),
-                'curl -Sk -X POST {}/query/configure'.format(HISTORY),
+                'curl -Sk -X GET --retry 60 --retry-delay 10 {}/_service/status > /dev/null'.format(HISTORY_URL),
+                'curl -Sk -X POST {}/query/configure'.format(HISTORY_URL),
             ]
 
         return shell_commands
 
     def action(self):
         check_config()
-        self.prev_version = StrictVersion(getenv('BREWBLOX_CFG_VERSION', '0.0.0'))
+        self.prev_version = StrictVersion(getenv(CFG_VERSION_KEY, '0.0.0'))
 
         if self.prev_version.version == (0, 0, 0):
             print('This configuration was never set up. Please run brewblox-ctl setup first')
@@ -55,6 +60,11 @@ class MigrateCommand(Command):
         if self.prev_version == StrictVersion(CURRENT_VERSION):
             print('Your system already is running the latest version ({})'.format(CURRENT_VERSION))
             return
+
+        if self.prev_version >= StrictVersion(CURRENT_VERSION):
+            print('Your system is running a version later than the latest. ' +
+                  'Please report a bug, or stop messing with the timeline.')
+            raise SystemExit(1)
 
         shell_commands = [
             '{}docker-compose down'.format(self.optsudo),
@@ -70,7 +80,7 @@ class MigrateCommand(Command):
         shell_commands += self.upped_commands()
 
         shell_commands += [
-            '{} -m dotenv.cli --quote never set BREWBLOX_CFG_VERSION {}'.format(sys.executable, CURRENT_VERSION),
+            '{} -m dotenv.cli --quote never set {} {}'.format(PY, CFG_VERSION_KEY, CURRENT_VERSION),
         ]
 
         self.run_all(shell_commands)
